+++ nmnb-application/src/main/kotlin/nmnb/application/global/infrastructure/external/KakaoOAuthClient.kt
@@ -1,4 +1,4 @@
-package nmnb.application.global.infrastructure.oauth
+package nmnb.application.global.infrastructure.external
 
 import com.fasterxml.jackson.databind.ObjectMapper
 import nmnb.application.global.auth.service.dto.KakaoProfile
