+++ nmnb-application/src/test/kotlin/nmnb/application/ControllerTestSupport.kt
@@ -9,8 +9,8 @@ import nmnb.application.domain.user.controller.UserController
 import nmnb.application.domain.user.service.UserService
 import nmnb.application.global.auth.generator.AuthUserArgumentResolver
 import nmnb.application.global.auth.service.AuthService
-import nmnb.application.global.auth.utils.JwtTokenProvider
 import nmnb.application.global.config.SecurityConfig
+import nmnb.application.global.infrastructure.security.JwtTokenProvider
 import nmnb.domain.user.repository.UserRepository
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest
