+++ nmnb-webflux/src/test/kotlin/nmnb/webflux/global/common/infrastructure/S3ServiceTest.kt
@@ -3,7 +3,7 @@ package nmnb.webflux.global.common.infrastructure
 import kotlinx.coroutines.test.runTest
 import nmnb.common.properties.S3Properties
 import nmnb.webflux.IntegrationTestSupport
-import nmnb.webflux.global.common.infrastructure.s3.S3Service
+import nmnb.webflux.global.common.infrastructure.external.S3Service
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.DisplayName
 import org.junit.jupiter.api.Test
